react router setup for page switching
