+++ client/package.json
@@ -9,6 +9,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.8.1",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
@@ -38,8 +39,8 @@
   },
   "devDependencies": {
     "@types/express": "^4.17.16",
-    "@types/node": "^18.11.18",
     "@types/jest": "^27.5.2",
+    "@types/node": "^18.11.18",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
     "typescript": "^4.9.4"
