+++ client/src/layouts/rootLayout.tsx
@@ -1,0 +1,15 @@
+import {NavLink, Outlet} from 'react-router-dom'
+
+export default function RootLayout() {
+
+    return (
+        <div className='root-layout'>
+
+
+        
+                <Outlet />
+           
+
+        </div>
+    )
+}
